test: Merge xml builder tests
