+++ tests/PhpPact/Xml/XmlBuilderTest.php
@@ -2,119 +2,149 @@
 
 namespace PhpPactTest\Xml;
 
+use PhpPact\Consumer\Matcher\Matcher;
 use PhpPact\Xml\Exception\InvalidXmlElementException;
 use PHPUnit\Framework\TestCase;
 use PhpPact\Xml\XmlBuilder;
 
 class XmlBuilderTest extends TestCase
 {
+    private XmlBuilder $builder;
+    private Matcher $matcher;
+
+    protected function setUp(): void
+    {
+        parent::setUp();
+
+        $this->builder = new XmlBuilder('1.0', 'UTF-8');
+        $this->matcher = new Matcher();
+    }
+
     public function testJsonSerializeInvalidXmlElement(): void
     {
         $this->expectException(InvalidXmlElementException::class);
         $this->expectExceptionMessage("Xml element's name is required");
 
-        $builder = new XmlBuilder('1.0', 'UTF-8');
-
-        $builder
+        $this->builder
             ->root(
-                $builder->name('Root'),
-                $builder->add(),
+                $this->builder->name('Root'),
+                $this->builder->add(),
             )
         ;
 
-        json_encode($builder);
+        json_encode($this->builder);
     }
 
     public function testJsonSerialize(): void
     {
-        $builder = new XmlBuilder('1.0', 'UTF-8');
-
-        $builder
+        $this->builder
             ->root(
-                $builder->name('Root'),
-                $builder->add(
-                    $builder->name('First Child Second Element'),
-                    $builder->contentLike('Example Test')
-                ),
-                $builder->add(
-                    $builder->name('Second Parent'),
-                    $builder->add(
-                        $builder->name('Second child 1'),
-                        $builder->attribute('myAttr', 'Attr Value')
+                $this->builder->name('ns1:projects'),
+                $this->builder->attribute('id', '1234'),
+                $this->builder->attribute('xmlns:ns1', 'http://some.namespace/and/more/stuff'),
+                $this->builder->eachLike(
+                    $this->builder->examples(2),
+                    $this->builder->name('ns1:project'),
+                    $this->builder->attribute('id', $this->matcher->integerV3(1)),
+                    $this->builder->attribute('type', 'activity'),
+                    $this->builder->attribute('name', $this->matcher->string('Project 1')),
+                    $this->builder->attribute('due', $this->matcher->datetime("yyyy-MM-dd'T'HH:mm:ss.SZ", '2016-02-11T09:46:56.023Z')),
+                    $this->builder->contentLike('Project 1 description'),
+                    $this->builder->add(
+                        $this->builder->name('ns1:tasks'),
+                        $this->builder->eachLike(
+                            $this->builder->examples(5),
+                            $this->builder->name('ns1:task'),
+                            $this->builder->attribute('id', $this->matcher->integerV3(1)),
+                            $this->builder->attribute('name', $this->matcher->string('Task 1')),
+                            $this->builder->attribute('done', $this->matcher->boolean()),
+                            $this->builder->contentLike('Task 1 description'),
+                        ),
                     ),
-                    $builder->add(
-                        $builder->name('Second child 2'),
-                        $builder->content('Test')
-                    ),
-                    $builder->add(
-                        $builder->name('Third Parent'),
-                        $builder->eachLike(
-                            $builder->name('Child')
-                        )
-                    ),
                 ),
-                $builder->add(
-                    $builder->name('First Child Third Element'),
-                ),
-            )
-        ;
+            );
 
         $expectedArray = [
             'version' => '1.0',
             'charset' => 'UTF-8',
             'root' => [
-                'name' => 'Root',
+                'name' => 'ns1:projects',
                 'children' => [
                     [
-                        'name' => 'FirstChildSecondElement',
-                        'children' => [
-                            [
-                                'content' => 'Example Test',
-                                'matcher' => ['pact:matcher:type' => 'type'],
-                            ],
-                        ],
-                        'attributes' => [],
-                    ],
-                    [
-                        'name' => 'SecondParent',
-                        'children' => [
-                            [
-                                'name' => 'Secondchild1',
-                                'children' => [],
-                                'attributes' => ['myAttr' => 'Attr Value'],
-                            ],
-                            [
-                                'name' => 'Secondchild2',
-                                'children' => [['content' => 'Test']],
-                                'attributes' => [],
-                            ],
-                            [
-                                'name' => 'ThirdParent',
-                                'children' => [
-                                    [
-                                        'pact:matcher:type' => 'type',
-                                        'value' => [
-                                            'name' => 'Child',
-                                            'children' => [],
-                                            'attributes' => [],
+                        'pact:matcher:type' => 'type',
+                        'value' => [
+                            'name' => 'ns1:project',
+                            'children' => [
+                                [
+                                    'name' => 'ns1:tasks',
+                                    'children' => [
+                                        [
+                                            'pact:matcher:type' => 'type',
+                                            'value' => [
+                                                'name' => 'ns1:task',
+                                                'children' => [
+                                                    [
+                                                        'content' =>
+                                                            'Task 1 description',
+                                                        'matcher' => [
+                                                            'pact:matcher:type' =>
+                                                                'type',
+                                                        ],
+                                                    ],
+                                                ],
+                                                'attributes' => [
+                                                    'id' => [
+                                                        'pact:matcher:type' =>
+                                                            'integer',
+                                                        'value' => 1,
+                                                    ],
+                                                    'name' => [
+                                                        'pact:matcher:type' => 'type',
+                                                        'value' => 'Task 1',
+                                                    ],
+                                                    'done' => [
+                                                        'pact:matcher:type' => 'type',
+                                                        'value' => true,
+                                                    ],
+                                                ],
+                                            ],
+                                            'examples' => 5,
                                         ],
                                     ],
+                                    'attributes' => [],
                                 ],
-                                'attributes' => [],
+                                [
+                                    'content' => 'Project 1 description',
+                                    'matcher' => ['pact:matcher:type' => 'type'],
+                                ],
+                            ],
+                            'attributes' => [
+                                'id' => [
+                                    'pact:matcher:type' => 'integer',
+                                    'value' => 1,
+                                ],
+                                'type' => 'activity',
+                                'name' => [
+                                    'pact:matcher:type' => 'type',
+                                    'value' => 'Project 1',
+                                ],
+                                'due' => [
+                                    'pact:matcher:type' => 'datetime',
+                                    'format' => "yyyy-MM-dd'T'HH:mm:ss.SZ",
+                                    'value' => '2016-02-11T09:46:56.023Z',
+                                ],
                             ],
                         ],
-                        'attributes' => [],
-                    ],
-                    [
-                        'name' => 'FirstChildThirdElement',
-                        'children' => [],
-                        'attributes' => [],
+                        'examples' => 2,
                     ],
                 ],
-                'attributes' => [],
+                'attributes' => [
+                    'id' => '1234',
+                    'xmlns:ns1' => 'http://some.namespace/and/more/stuff',
+                ],
             ],
         ];
 
-        $this->assertSame(json_encode($expectedArray), json_encode($builder));
+        $this->assertSame(json_encode($expectedArray), json_encode($this->builder));
     }
 }
